requirements.txt
pandas
plotly
prophet
pystan==2.19.1.1
pillow
kaleido

